Add/edit admin dashboard components for full functionality
Dashboard.js is the containing component, within which lies
NewRound, AddUser, and AddExercise.  NewRound is designed to initiate
a new round manually (selecting a name, and an exercise from the
existing database of exercises).  AddUser manually adds a user, and
if they're added in a later round, enough equivalent 0's are placed
in that user's scores array in the database to keep the array
of the same length as users who have existed from the start.
AddExercise allows for creating a new exercise in the database, which
can then be used as the chosen exercise for a later round.